+++ packages/cli/templates/open-next.config.ts
@@ -1,12 +1,12 @@
-// @ts-nocheck
 import { defineCloudflareConfig } from '@opennextjs/cloudflare';
 import { purgeCache } from '@opennextjs/cloudflare/overrides/cache-purge/index';
 import kvIncrementalCache from '@opennextjs/cloudflare/overrides/incremental-cache/kv-incremental-cache';
 import doQueue from '@opennextjs/cloudflare/overrides/queue/do-queue';
 import queueCache from '@opennextjs/cloudflare/overrides/queue/queue-cache';
 import doShardedTagCache from '@opennextjs/cloudflare/overrides/tag-cache/do-sharded-tag-cache';
+import { type OpenNextConfig } from '@opennextjs/cloudflare';
 
-export default defineCloudflareConfig({
+const cloudflareConfig = defineCloudflareConfig({
   incrementalCache: kvIncrementalCache,
   queue: queueCache(doQueue, {
     regionalCacheTtlSec: 5,
@@ -27,3 +27,10 @@ export default defineCloudflareConfig({
   enableCacheInterception: false,
   cachePurge: purgeCache({ type: 'durableObject' }),
 });
+
+const config: OpenNextConfig = {
+  buildCommand: 'node_modules/.bin/next build',
+  ...cloudflareConfig,
+};
+
+export default config;